+++ catwalk/models/rank_classification.py
@@ -205,6 +205,8 @@ class TrainableRankClassificationModel(TrainableModel):
 
 @Model.register("rc::encoder_decoder")
 class EncoderDecoderRCModel(RankClassificationModel):
+    VERSION = RankClassificationModel.VERSION + "002spt"
+
     @classmethod
     def _make_model(cls, pretrained_model_name_or_path: str, *, make_copy: bool = False, **kwargs) -> T5ForConditionalGeneration:
         return cached_transformers.get(AutoModelForSeq2SeqLM, pretrained_model_name_or_path, make_copy=make_copy, **kwargs)
@@ -274,6 +276,13 @@ class DecoderOnlyRCModel(RankClassificationModel):
     def _make_model(cls, pretrained_model_name_or_path: str, *, make_copy: bool = False, **kwargs) -> GPT2LMHeadModel:
         return cached_transformers.get(AutoModelForCausalLM, pretrained_model_name_or_path, make_copy=make_copy, **kwargs)
 
+    @staticmethod
+    def _prefix_with_space(s: str) -> str:
+        if not s.startswith(' '):
+            return f" {s}"
+        else:
+            return s
+
     def _run_loglikelihood(
         self,
         tuples: Sequence[Tuple[str, str]],
@@ -281,8 +290,8 @@ class DecoderOnlyRCModel(RankClassificationModel):
         tokenizer: _Tokenizer,
         batch_size: int = 32,
     ) -> Sequence[float]:
-        tokenized_contexts = tokenizer([t[0] for t in tuples])
-        tokenized_continuations = tokenizer([t[1] for t in tuples])
+        tokenized_contexts = tokenizer([t[0] for t in tuples], add_special_tokens=False)
+        tokenized_continuations = tokenizer([self._prefix_with_space(t[1]) for t in tuples], add_special_tokens=False)
 
         # transpose the token ids so we can access them one instance at a time
         cc_pairs: List[Dict[str, Tuple[torch.Tensor, torch.Tensor]]] = []